+++ CHANGELOG.md
@@ -7,9 +7,13 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ## Unreleased
 
 ### New
+- new `generateName` attribute on DeploymentManifest for dynamic creation of the deployment `name`
+
 ### Changes
 
 ### Fixes
+- unable to destroy when module sourced from remote git repository 
+
 
 ## v2.2.1 (2022-10-25)